XfsRadioButton.c
The additional file needed to recreate my twenty button X Radio Button project. Compile it with the line:
gcc -o XRadioButton XfsRadioButton.c XRadioButton.c -lX11 -I.
Use camel case if you like, that does not matter. What matters is you use the names you gave the files in the compile command exactly and the flags: hypen "l" as in lunch X11 <space> hypen I as in the Roman Numeral One dot <.> DOES matter. Run it from a terminal and it is loads of summer time fun for the kids!